Scrawford dev (#7)
* Simulation works.

* Ok, sim doesn't work. Help!

* Sim runs without issues. Need to check error logic.

* TB successful for direct mapped.

* Two-way testbench still has errors

Testbench  passes for direct mapped. For NUM_SETS = 2, I'm not sure why we're reporting hits as often as we are which causes data to (mostly) write into set number 1.

* Updating testbench.

Testbench passes for various direct mapped cache sizes. Fails for 2-way set associative. Probably an error in the actual cache, not the testbenh.